% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatSS.R
\name{formatSS}
\alias{formatSS}
\title{Format GWAS results}
\usage{
formatSS(
  gwasfile = NULL,
  outstem = NULL,
  NameCol = NULL,
  A1Col = NULL,
  A2Col = NULL,
  ZCol = NULL,
  EffectCol = NULL,
  SECol = NULL,
  PCol = NULL,
  nCol = NULL,
  ncasesCol = NULL,
  ncontrolsCol = NULL,
  fixedn = NULL,
  FreqCol = NULL,
  headerRows = NULL
)
}
\arguments{
\item{gwasfile}{The name of the file containing GWAS results}
}
\description{
Format GWAS results
}
