% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megaprs_main.R
\name{megaprs}
\alias{megaprs}
\title{Estimate predictor effect sizes}
\usage{
megaprs(
  corstem = NULL,
  ssfile = NULL,
  outfile = NULL,
  power = -0.25,
  shrink = 0.9,
  jackknife = FALSE,
  perpredhers = NULL,
  onechr = NULL,
  extractfile = NULL
)
}
\arguments{
\item{corstem}{The stem of the correlations files (e.g., if you have files called example.cors.root, example.cors.windows, etc, then use "corstem=example")}

\item{ssfile}{The name of file containing GWAS summary statistics in LDAK format (which you either created yourself, or using the function formatSS)}

\item{outfile}{The name of the file where effect sizes will be saved}

\item{power}{Defines the relationship between per-predictor heritability and MAF (Default value is -0.25)}

\item{shrink}{How much to shrink predictor-predictor correlations (Default value is 0.8)}

\item{jackknife}{Whether to estimate standard deviations via jackknifing (Default value FALSE)}

\item{perpredhers}{(Optional) Used to provide per-predictor heritabilities (the file should have two columns, providing predictor names and heritabilities)}

\item{onechr}{(Optional) Used to restrict to predictors on a single chromosome}

\item{extractfile}{(Optional) Used to restrict to predictors in the specified file#' @export}
}
\description{
Estimate predictor effect sizes
}
\examples{
megaprs(corstem="example", ssfile="gwas.summaries", outfile="output.effects")
}
